+++ src/router.js
@@ -10,7 +10,7 @@ import Footer from "./components/Footer";
 import Loading from "./components/Loading";
 import { LogIn } from "./components/Login";
 import { Main } from "./components/Main";
-import ContactInfoPage from "./components/SignUp/ContactInfoPage";
+import ContactInfoPage from "./components/Profile/ContactInfoPage";
 import TriplersPage from "./components/Triplers/TriplersPage";
 import TriplersAdd from "./components/Triplers/AddTripler";
 import ConfirmPage from "./components/Triplers/ConfirmPage";
